[US_MO][Tasks] Add clients in CSC districts to candidate population (Recidiviz/recidiviz-data#53454)
## Description of the change

Added clients in CSC districts back into the candidate population for MO
Tasks, as we don't want to filter out non-residential clients in those
districts from the tool. We'll still need to go back and filter out CSC
residents, but we don't yet know how to do that, so there's a TODO for
it.

To load views to sandbox for validation:
```
python -m recidiviz.tools.load_views_to_sandbox \
--sandbox_dataset_prefix mhilderbran_us_mo_tasks_csc \
--prompt auto \
--load_up_to_addresses workflows_views.us_mo_supervision_tasks_record,observations__person_event.supervision_contact_due
```

We now see more clients with eligible tasks in the tasks record, as
expected:
```
WITH main AS (
  SELECT
    state_code,
    JSON_VALUE(task, '$.type') AS type,
    COUNT(DISTINCT person_id) AS n_clients_with_task_main,
  FROM `workflows_views.us_mo_supervision_tasks_record_materialized`,
  UNNEST(tasks) AS task
  GROUP BY 1, 2
),
sandbox AS (
  SELECT
    state_code,
    JSON_VALUE(task, '$.type') AS type,
    COUNT(DISTINCT person_id) AS n_clients_with_task_sandbox,
  FROM `mhilderbran_us_mo_tasks_csc_workflows_views.us_mo_supervision_tasks_record_materialized`,
  UNNEST(tasks) AS task
  GROUP BY 1, 2
)
SELECT *
FROM main
FULL OUTER JOIN sandbox
  USING (state_code, type)
ORDER BY state_code, type
```

## Related issues

Relates to Recidiviz/recidiviz-data#50537

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### ⚠️ Changing an opportunity record?
Please [consult this
document](https://www.notion.so/recidiviz/How-to-make-a-change-to-the-Opportunity-Record-1d67889f4d19807f8109f9f2f007dde3)
first to see if additional changes might be required in the admin panel
or a Zod schema. You can confirm with Polaris if unsure.

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 17c890dc3070db7e0048133bbdc49f5b1017f44a