[US_IA][Workflows] Modify IA SLD completion event (Recidiviz/recidiviz-data#53437)
## Description of the change

Update the completion event used for IA SLD to only include the L2->L1
transitions since that's the only group we are currently surfacing in
the tool. Tested the change in a sandbox and confirmed the IA SLD
transitions are properly getting counted towards the new completion
event type.

## Related issues

Relates to Recidiviz/recidiviz-data#48523

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1b18d40c250b5a8cb497fdfc04286f9f5de30a9a