[Auth] Add jii_pages and jii_permissions to roster/auth tables and endpoints (Recidiviz/recidiviz-data#53439)
## Description of the change

We want to control staff access to jii though the normal admin panel
views the same way we do it for the dashboard app. This PR adds two new
columns: `allowed_apps` and `jii_permissions`, adds a migration, and
updates the auth endpoints to handle these new fields.

The `allowed_apps` field will be used to gate overall user access to the
staff and jii apps. For now, I will only check it in the JII auth flow.
Once we know it works there, we can add support for it in the dashboards
auth flow. This field is not intended to be passed to the fronted app
itself in `app_metadata` or the like.

Tested via unit tests.

I will then follow up with updates to the admin panel to show and edit
these fields.

Finally, we will update the auth flow in the JII app to call these
endpoints when a staff user logs in.

More details
[here](https://www.notion.so/recidiviz/Supporting-staff-access-to-JII-app-via-admin-panel-29b7889f4d19800e9896d50609b26d2c?showMoveTo=true&saveParent=true).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a31a0cc7b783310b891a9ed5955e3a47f76d6aa6