[US_AZ] State sentence group length & incarceration period improvements (Recidiviz/recidiviz-data#53531)
## Description of the change

This makes three improvements that @agaidus identified during his
validation of release dates in AZ.

1. if LIFE_OR_DEATH_ID is not null, set the date to NULL
2. if the _ML flag is set for the sentence as a whole, use the _ML field
for every date
3. bump the allowable release dates to anything before the year 2700
(incorporating anna's change in
https://github.com/Recidiviz/recidiviz-data/pull/53583)

This PR _also_ fixes a bug in incarceration periods.

I incorporated internal movements into incarceration periods in
https://github.com/Recidiviz/recidiviz-data/pull/51405, but I did not
adjust the filter that ensures we ignore administrative updates that
happened when a person was not actually incarcerated. This PR makes it
so that if a person was "routinely unassigned from a bed" 10 years after
they were released from prison, we discount that movement as being
administrative and not legitimate. This reduces the active prison
population by 604, and I validated 20 of the people who have newly
closed periods against ACIS. The new logic is more correct than the
previous logic - none of those people should have had open incarceration
periods.

[Successful ingest
pipeline](https://console.cloud.google.com/dataflow/jobs/us-central1/2025-11-21_10_10_58-11905350529926870972?project=recidiviz-staging).
Loaded end to end sandbox for validation of release dates with prefix
`elise_dates`.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#53490
Closes Recidiviz/recidiviz-data#53519

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### ⚠️ Changing an opportunity record?
Please [consult this
document](https://www.notion.so/recidiviz/How-to-make-a-change-to-the-Opportunity-Record-1d67889f4d19807f8109f9f2f007dde3)
first to see if additional changes might be required in the admin panel
or a Zod schema. You can confirm with Polaris if unsure.

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8987ca6a951bc5b95a8d9493f9c326f3e5b11e85