[US_IX] Tasks: Create a state-specific case type supervision level view (Recidiviz/recidiviz-data#53941)
## Description of the change

[US_IX] Tasks: Create a state-specific case type supervision level view

I uploaded to a sandbox and it's all looking ok! There are no changes in
the population eligible for a task because we still need the
`compartment_level_1 != 'INTERNAL_UNKNOWN'`, but that needs to happen
during ingest.

```
 python -m recidiviz.tools.load_views_to_sandbox \
        --project_id recidiviz-staging\
        --sandbox_dataset_prefix hsalas_ass\
        --state_code_filter US_IX manual\
        --dataset_ids_to_load tasks_views,task_eligibility_criteria_us_ix,compliance_task_eligibility_spans_us_ix,task_eligibility_criteria_general,reference_views\
        --update_descendants False
```

```
WITH sandbox AS (
  SELECT
    task_name, COUNT(*) AS count_sandbox
  FROM recidiviz-staging.hsalas_compliance_task_eligibility_spans_us_ix.all_compliance_task_eligibility_spans_materialized
  WHERE CURRENT_DATE BETWEEN start_date AND IFNULL(end_date, '9999-12-31')
    AND is_eligible
  GROUP BY 1
),

staging AS (
  SELECT
    task_name, COUNT(*) AS count_staging
  FROM recidiviz-staging.compliance_task_eligibility_spans_us_ix.all_compliance_task_eligibility_spans_materialized
  WHERE CURRENT_DATE BETWEEN start_date AND IFNULL(end_date, '9999-12-31')
    AND is_eligible
  GROUP BY 1
)

SELECT *, (count_sandbox-count_staging)/count_staging*100 AS percentage_change
FROM sandbox
LEFT JOIN staging
  USING(task_name)
```
## Related issues

Closes
[Recidiviz/recidiviz-data#53577](https://github.com/Recidiviz/recidiviz-data/issues/53577)

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### ⚠️ Changing an opportunity record?
Please [consult this
document](https://www.notion.so/recidiviz/How-to-make-a-change-to-the-Opportunity-Record-1d67889f4d19807f8109f9f2f007dde3)
first to see if additional changes might be required in the admin panel
or a Zod schema. You can confirm with Polaris if unsure.

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a85f6d5be780aa0c507f603d6b2049185ff0267a