[US_NY] Remove ethnicity mapping from state_person (Recidiviz/recidiviz-data#54095)
## Description of the change

For the next version of NY Pathways, we need to be able to break
populations down by race, which is getting pulled in from
person_demographics. However, person_demographics only has a field for
prioritized_race_or_ethnicity and not race specifically. By removing the
mapping for state_person ethnicity, we can ensure that
prioritized_race_or_ethnicity will always be directly hydrated with
state_person race. However, for future versions, we'll need to figure
out a way to get both race and ethnicity downstream separately, at which
point we'll want to revert this change (ticket here: Recidiviz/recidiviz-data#54092).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### ⚠️ Changing an opportunity record?
Please [consult this
document](https://www.notion.so/recidiviz/How-to-make-a-change-to-the-Opportunity-Record-1d67889f4d19807f8109f9f2f007dde3)
first to see if additional changes might be required in the admin panel
or a Zod schema. You can confirm with Polaris if unsure.

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8de7d174ac8297c2ea24ea7ed0e9babd4a81e7f3