[deploy] Few fixes to the deploy script (Recidiviz/recidiviz-data#53939)
## Description of the change

- add gcloud auth check to predeploy checks since it is needed in looker
deploy and elsewhere
- fail loudly if the gcloud command fails in `update_deployment_status`.
maybe we don't actually care if this step fails? but since it is only
run before/after the deploy i dont think failing at this step will leave
us in a bad state and i think its better to fail than to just have to
notice that this step is not working
- checkout release tag before `verify_can_deploy` steps. i noticed in
the most recent deploy log that we passed the pipenv sync check in the
predeploy check then later got a message about pipenv being out of sync
because we ran the check while we still had `main` checked out.

full context
https://recidiviz.slack.com/archives/CJDCVR2AY/p1764102721214049

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### ⚠️ Changing an opportunity record?
Please [consult this
document](https://www.notion.so/recidiviz/How-to-make-a-change-to-the-Opportunity-Record-1d67889f4d19807f8109f9f2f007dde3)
first to see if additional changes might be required in the admin panel
or a Zod schema. You can confirm with Polaris if unsure.

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 65359eec896c67034e7117e1fc3623f3cc426b5a