[raw data] Add file chunking metadata classes (Recidiviz/recidiviz-data#54150)
## Description of the change

during sftp file transfer, we want to be able to enforce that all
expected files are present before uploading them to the ingest bucket.
in order to do that, we need to know how many chunks are expected if a
file is chunked. the way we currently handle chunked files is by
defining state-specific delegates that define a custom
`coalesce_chunked_files` function that groups gcs files into a single
logical bq file if that state has chunked files. this leaves us unable
to reason about which files are chunked outside of the context of that
function.

instead, introduce new classes to define a file's chunking behavior over
time. see the new `coalesce_chunked_files` logic and state-specific
configurations that are intended to replace the state-specific delegates
in https://github.com/Recidiviz/recidiviz-data/pull/54148. we can then
also use the state-specific chunking metadata histories during sftp to
determine the number of chunks that should be present and enforce that
we aren't trying to enforce expected files present for states that have
chunked files with an unknown number of chunks.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#53517

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### ⚠️ Changing an opportunity record?
Please [consult this
document](https://www.notion.so/recidiviz/How-to-make-a-change-to-the-Opportunity-Record-1d67889f4d19807f8109f9f2f007dde3)
first to see if additional changes might be required in the admin panel
or a Zod schema. You can confirm with Polaris if unsure.

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 53bc1c03f65e96ef81511817cc225cb9bcafbab1